+++ src/main/core/content/entity/BaseBlockEntity.java
@@ -4,7 +4,7 @@ import core.content.blocks.Block;
 
 public abstract class BaseBlockEntity<B extends Block> implements BlockEntity {
 
-    protected int x, y;
+    protected short x, y;
     protected final B block;
 
     protected BaseBlockEntity(B block) {
@@ -13,10 +13,10 @@ public abstract class BaseBlockEntity<B extends Block> implements BlockEntity {
 
     public final B type() { return block; }
 
-    public int x() { return x; }
-    public int y() { return y; }
+    public short x() { return x; }
+    public short y() { return y; }
 
-    public void setPosition(int x, int y) {
+    public void setPosition(short x, short y) {
         this.x = x;
         this.y = y;
     }
